pyproject.toml
[tool.poetry]
name = "send-messages-bot"
version = "0.1.0"
description = ""
authors = ["Enmanuel Silva <enmanuel.silva@spotcloud.io>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.0.0"
python-dotenv = "^0.20.0"
pydantic = "^1.9.2"
psycopg2-binary = "^2.9.3"
sqlmodel = "^0.0.6"
alembic = "^1.8.1"
twilio = "^7.13.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
commitizen = "^2.27.1"
pre-commit = "2.19.0"
black = "22.6.0"
flake8 = "4.0.1"
bandit = "1.7.4"
isort = "5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_files = [
    "src/__version__.py",
    "pyproject.toml:version"
]
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]
